backend/requirements.txt
aiofiles==23.2.1
blinker==1.7.0
click==8.1.7
exceptiongroup==1.1.3
Flask==3.0.0
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.15.0
hyperframe==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
pandas==2.1.3
priority==2.0.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
Quart==0.19.3
six==1.16.0
SQLAlchemy==2.0.23
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==3.0.1
wsproto==1.2.0
